Fix helmrepo reconciler unfinished worker bug
Although all the APIs had interval as a required field, when tests
objects were created, they had the zero value of interval, which the API
server accepts. A zero interval value results in the test objects to
reconcile only once when they are created and never reconcile again
unless there's an update to the object. Most of the tests worked with
this behavior.

With HelmRepository removing the interval requirement and adding an
internal default, all the HelmRepository objects created in the tests
without any interval have a default interval value which results in
objects to reconcile automatically if they are not cleaned up after
running tests. TestHelmRepositoryReconciler_InMemoryCaching creates
HelmRepository but doesn't delete it at the end. This leads to a
reconciliation of HelmRepository outside of the test in the envtest
environment. It just happened to be that the reconciliation time matches
with the end of test time. At the end of the test run, the reconcilers
receive shutdown signal and any test server, like helmrepository server,
are stopped. A HelmRepository reconciliation triggered just before the
shutdown signal gets stuck in the reconciliation. HelmRepository can't
download the index as the test index server has stopped and hangs for
some time. The HelmRepository reconciler worker remains in active state,
unlike other reconciler workers that shut down, resulting in the test to
timeout at the end.

The is fixed by deleting the HelmRepository object created in
TestHelmRepositoryReconciler_InMemoryCaching at the end of the test
similar to other tests.

Signed-off-by: Sunny <darkowlzz@protonmail.com>